+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-04 10:20:21
+
+### Changed Files
+- `src/hooks/targets/useHistoricalAverages.ts`
+- `src/services/targets/currentMonthMetricsService.ts`
+
+### Statistics
+```
+ 2 files changed, 54 insertions(+), 12 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-04 09:59:23
 
 ### Changed Files